requirements.txt
streamlit>=1.32
simple-salesforce>=1.12
plotly>=5.18
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0
